Add files via upload
Question-Can We Update the value of a key in dictionary ?If yes,how?
Answer-The update method updates the dictionary with the elements from the another dictionary object or from an iterable of key/value pairs. update method adds element to the dictionary if the key is not in the dictionary. If the key is in the dictionary, it updates the key with the new value.